backend/requirements.txt
fastapi<=0.95.2
pydantic<=1.10.8
uvicorn<=0.22.0
python-dotenv<=1.0.0
FastAPI-SQLAlchemy<=0.2.1
typing<=3.7.4.3
SQLAlchemy<=2.0.15
